+++ index.js
@@ -107,6 +107,10 @@ function devMode(enabled) {
   atlas.setDevMode(enabled);
 }
 
+function analyzeId(name, tags) {
+  return atlas.analyzeId(name, tags);
+}
+
 let scope = function(commonTags) {
   let bucketArgs = function(name) {
     let tags, bucketFunction;
@@ -127,6 +131,7 @@ let scope = function(commonTags) {
     stop: stopAtlas,
     setDevMode: devMode,
     getDebugInfo: debugInfo,
+    analyzeId: analyzeId,
     counter: (name, tags) => atlas.counter(
       name, Object.assign({}, commonTags, tags)),
     dcounter: (name, tags) => atlas.dcounter(